Move approve button to components folder
