+++ ui/components/Swap/ApproveQuoteButton.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { useTranslation } from "react-i18next"
-import SharedButton from "../components/Shared/SharedButton"
+import SharedButton from "../Shared/SharedButton"
 
 export default function ApproveQuoteBtn({
   isApprovalInProgress,
